requirements.txt
Flask
torch
Pillow
numpy
opencv-python
base64
